Authentication

+++ backend/src/Todo.Api/Slices/Todo/RemoveToDo.cs
@@ -1,6 +1,7 @@
 using FluentValidation;
 using MediatR;
 using MongoDB.Bson;
+using Todo.Api.Authentication;
 using Todo.Api.Domain.Infrastructure;
 using Todo.Api.Domain.Todo;
 
@@ -32,15 +33,27 @@ public abstract class RemoveToDo
     public class RemoveToDoHandler : IRequestHandler<RemoveToDoRequest, Response>
     {
         private readonly IDocumentRepository<ToDoDocument> _documentRepository;
+        private readonly IUserProfileAccessor _userProfileAccessor;
 
-        public RemoveToDoHandler(IDocumentRepository<ToDoDocument> documentRepository)
+        public RemoveToDoHandler(IDocumentRepository<ToDoDocument> documentRepository, IUserProfileAccessor userProfileAccessor)
         {
             _documentRepository = documentRepository;
+            _userProfileAccessor = userProfileAccessor;
         }
     
         public async Task<Response> Handle(RemoveToDoRequest request, CancellationToken cancellationToken)
         {
+            var userId = _userProfileAccessor.Subject;
+            
+            var todo = await _documentRepository.FindOneToDoAsync(request.Id!);
+            
+            if (userId != todo!.UserId)
+            {
+                throw new UnauthorizedAccessException("You do not have access to this todo");
+            }
+            
             await _documentRepository.RemoveToDoAsync(request.Id);
+            
             return new Response();
         }
     }
